backend/requirements.txt
fastapi
uvicorn
pandas
pydantic
python-multipart
google-generativeai
python-dotenv
requests